Should work, but doesn't.  need to debug but no time right now.
