+++ OpenPNM/Network/__init__.py
@@ -31,6 +31,7 @@ Classes
 
 """
 
+from . import tools
 from .__GenericNetwork__ import GenericNetwork
 from .__Cubic__ import Cubic
 from .__Delaunay__ import Delaunay
@@ -39,3 +40,4 @@ from .__MatFile__ import MatFile
 from .__TestNet__ import TestNet
 from .__Empty__ import Empty
 from . import models
+
